requirements.txt
3==0.0.0
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
confusable-homoglyphs==3.2.0
Django==1.11
django-bootstrap==0.2.4
django-bootstrap3==12.0.3
django-registration==2.4.1
django-tinymce==2.8.0
djangorestframework==3.11.0
idna==2.8
Pillow==7.0.0
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
pyuploadcare==2.6.0
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
